Update codeOfConduct.md
### Stirling Coding Club CoC
### The assignment below is taken from (https://mozillascience.github.io/study-group-orientation/1.3-writing-a-coc.html)
1.3 Assignment: Writing a Code of Conduct
Mozilla Study Groups should be welcoming environments for all participants, regardless of their gender, ethnic background, sexual orientation, age, skill level, preference for text editor or coding langauge, etc. Study Groups should get everyone learning together!

Take some time to think the learning environment you'd like to create by writing a Code of Conduct, also known sometimes as Participation Guidelines. This document describes your ideal Group culture, and defines social norms and responsibilities for members of your Group. The guidelines apply to in person interactions and those that happen online, in forums, chat, via email and in your Study Group repositories.

It may seem unnecessary to have these, but when working in a collaborative group, it is always possible (and maybe inevitable) that tensions may arise! You may have seen this in your own research labs or classes. A Code of Conduct puts people in the right mindset to work together.

A Code of Conduct should:

Establish the positive behaviors encouraged by the community
Make clear which behaviors are unacceptable and discouraged
Define a process by which any problems will be handled.
Write your Code of Conduct
Brainstorm. Reflect on the following questions.
What core words would you associate with the Group you hope to create? These could be values, ideals, or characteristics, and behaviors. Try to keep these to one word answers, if possible. (Examples: friendly, productive, relaxed sharing, helping, asking questions, etc).
What behaviors to discourage? Be specific enough to be useful here. (interrupting, put-downs, unwanted physical contact, sexist comments)
What should someone do if they have an issue or problem with a behavior in the group? This may be as simple as speaking to the group leader, or sending a email or text. Be explicit in these steps. It’s OK (and encouraged) to have a few different options for reporting issues.
What consequences are there for violating the code? It’s important the consequences are real and appropriate to the situation. People might be warned, or asked to leave the session, or they may be asked to leave the Group.
Who decides what does and does not violate the code? What's an example of how this might be done? As Group Lead, you're primarily responsible for making sure the group goes well, but as you add members, you might ask another trusted member to help you make these decisions.

Refine and Remix Using the information you’ve collected in the brainstorm above, write your own Code of Conduct. Remember, Codes of Conduct are about creating a welcoming and safe, and postive environment as much as they are about discouraging bad behavior. As you write be sure to highight the desired values of the group. You're also welcome to refer to or remix our Code of Conduct.

Share and Discuss If you already have a few members, share these and discuss them, and revise as needed.